feat: Document innovations and fix platform leaks
- Updated `bolt.md` with proposed innovations including NeuralFX, TurboSampler, and HyperGraph.
- Fixed `audit_codebase.py` to ignore `= delete` syntax.
- Refactored `AestraThreading` to move platform-specific implementations (MMCSS, ThreadPool) to a new `.cpp` file, eliminating `<windows.h>` from the public header.
- Updated `AudioEngine.h` to remove `<windows.h>` include.
- Added explicit `<windows.h>` includes with `// ALLOW_PLATFORM_INCLUDE` in implementation files where necessary.
- Updated `AestraCore/CMakeLists.txt` to include the new source file.
