+++ AestraAudio/src/Core/AudioEngine.cpp
@@ -16,6 +16,14 @@
 #include <immintrin.h> // AVX/SSE for high-performance mixing
 #include "miniaudio.h"
 
+#ifdef _WIN32
+#define WIN32_LEAN_AND_MEAN
+#ifndef NOMINMAX
+#define NOMINMAX
+#endif
+#include <windows.h> // ALLOW_PLATFORM_INCLUDE
+#endif
+
 // Denormal protection macros
 #define DISABLE_DENORMALS \
     int oldMXCSR = _mm_getcsr(); \
