+++ bolt.md
@@ -71,5 +71,12 @@ Move from a linear processing list to a DAG (Directed Acyclic Graph) task schedu
 - **Fix**: Replaced with `std::atomic<std::shared_ptr>` + Deferred Reclamation (GC).
 - **Violation**: `EffectChain` deleted operators (False Positive in audit, but good to know).
 
+### Platform Abstraction Leaks
+
+- **Violation**: `AestraThreading.h` included `<windows.h>` directly.
+- **Fix**: Moved implementation to `.cpp` with `// ALLOW_PLATFORM_INCLUDE`.
+- **Violation**: `AudioEngine.h` included `<windows.h>`.
+- **Fix**: Removed include, moved necessary parts to `.cpp`.
+
 ---
 *Signed: Bolt*
